addressed tab vs actions - hover/blue box still showing
