+++ src/components/content/conditions/patient-conditions-tabs.tsx
@@ -1,28 +1,20 @@
 import { Tab } from "@headlessui/react";
 import cx from "classnames";
-import { useAddConditionForm } from "./condition-hooks";
 import { FilterCollection } from "./patient-conditions-filters";
 import { Badge } from "@/components/core/badge";
-import { Toggle } from "@/components/core/toggle";
 import { ConditionModel } from "@/fhir/models";
 
-export type PatientConditionsActionsProps = {
-  hideAdd: boolean;
-  onToggleShowHistoric: () => void;
+export type PatientConditionsTabsProps = {
   collection: FilterCollection;
   otherConditions: ConditionModel[];
   onCollectionChange: (collection: FilterCollection) => void;
 };
 
-export function PatientConditionsActions({
-  hideAdd,
-  onToggleShowHistoric,
+export function PatientConditionsTabs({
   collection,
   otherConditions,
   onCollectionChange,
-}: PatientConditionsActionsProps) {
-  const showAddConditionForm = useAddConditionForm();
-
+}: PatientConditionsTabsProps) {
   function activeClass(collection2: FilterCollection) {
     return collection === collection2
       ? "ctw-text-content-black ctw-tab-underline ctw-relative"
@@ -34,22 +26,24 @@ export function PatientConditionsActions({
   ).length;
 
   return (
-    <>
-      <div className="ctw-justify-end ctw-space-x-2 ctw-border-0 ctw-border-b ctw-border-t ctw-border-solid ctw-border-divider-light ctw-py-5">
-        <div className="ctw-space-x-4">
-          <Tab.Group>
-            <Tab.List className="ctw-space-x-2">
-              <Tab
-                className={cx(
-                  activeClass("patient"),
-                  "ctw-tab-underline ctw-tab ctw-relative"
-                )}
-              >
+    <div className="ctw-justify-end ctw-space-x-2 ctw-border-0 ctw-border-b ctw-border-t ctw-border-solid ctw-border-divider-light ctw-py-5">
+      <div className="ctw-space-x-4">
+        <Tab.Group>
+          <Tab.List className="ctw-space-x-2">
+            <Tab
+              className={cx(
+                activeClass("patient"),
+                "ctw-tab-underline ctw-tab ctw-relative"
+              )}
+            >
+              {({ selected }) => (
                 <button
                   type="button"
                   className={cx(
                     activeClass("patient"),
-                    "ctw-tab-underline ctw-tab ctw-relative"
+                    selected
+                      ? "ctw-tab-underline-hover ctw-tab-underline ctw-tab ctw-relative"
+                      : "ctw-tab-underline-hover:focus-visible ctw-tab-underline ctw-tab ctw-relative"
                   )}
                   onClick={() => {
                     onCollectionChange("patient");
@@ -57,18 +51,22 @@ export function PatientConditionsActions({
                 >
                   Condition List
                 </button>
-              </Tab>
-              <Tab
-                className={cx(
-                  activeClass("other"),
-                  "ctw-tab-underline ctw-tab ctw-relative ctw-space-x-2"
-                )}
-              >
+              )}
+            </Tab>
+            <Tab
+              className={cx(
+                activeClass("other"),
+                "ctw-tab-underline ctw-tab ctw-relative ctw-space-x-2"
+              )}
+            >
+              {({ selected }) => (
                 <button
                   type="button"
                   className={cx(
                     activeClass("other"),
-                    "ctw-tab-underline ctw-tab ctw-relative ctw-space-x-2"
+                    selected
+                      ? "ctw-tab-underline ctw-tab-underline-hover ctw-tab ctw-relative ctw-space-x-2"
+                      : "ctw-tab-underline ctw-tab-underline-hover:focus-visible ctw-tab ctw-relative ctw-space-x-2"
                   )}
                   onClick={() => {
                     onCollectionChange("other");
@@ -77,29 +75,11 @@ export function PatientConditionsActions({
                   <span>Other Provider Records</span>
                   <Badge text={`${activeCount}`} color="primary" />
                 </button>
-              </Tab>
-            </Tab.List>
-          </Tab.Group>
-        </div>
+              )}
+            </Tab>
+          </Tab.List>
+        </Tab.Group>
       </div>
-      {!hideAdd ? (
-        <div className="ctw-flex ctw-items-center ctw-justify-end ctw-space-x-2 ctw-p-4">
-          <Toggle
-            name="historic"
-            text="Historic"
-            onChange={onToggleShowHistoric}
-          />
-          <button
-            type="button"
-            className="ctw-btn-primary"
-            onClick={() => showAddConditionForm()}
-          >
-            Add Condition
-          </button>
-        </div>
-      ) : (
-        <div className="ctw-p-4" />
-      )}
-    </>
+    </div>
   );
 }
