+++ src/components/content/conditions/patient-conditions.tsx
@@ -3,13 +3,14 @@ import "./patient-conditions.scss";
 import { useRef } from "react";
 import { useConditionHistory } from "../condition-history/conditions-history-drawer";
 import { filterOtherConditions } from "./helpers";
+import { PatientConditionsActions } from "./patient-conditions-actions";
 import { patientConditionsColumns } from "./patient-conditions-columns";
 import { useConditionFilters } from "./patient-conditions-filters";
 import {
   OtherProviderConditionHoverActions,
   PatientConditionHoverActions,
 } from "./patient-conditions-menu-actions";
-import { PatientConditionsActions } from "./patient-conditions-tabs";
+import { PatientConditionsTabs } from "./patient-conditions-tabs";
 import { FormEntry } from "@/components/core/form/drawer-form-with-fields";
 import { Table } from "@/components/core/table/table";
 import {
@@ -77,14 +78,16 @@ export function PatientConditions({
         <div className="ctw-text-xl ctw-font-medium ctw-text-content-black">
           Conditions
         </div>
+        <PatientConditionsTabs
+          otherConditions={otherConditions}
+          collection={filters.collection}
+          onCollectionChange={(collection) => updateFilters({ collection })}
+        />
         <PatientConditionsActions
           hideAdd={readOnly || filters.collection === "other"}
           onToggleShowHistoric={() =>
             updateFilters({ showHistoric: !filters.showHistoric })
           }
-          otherConditions={otherConditions}
-          collection={filters.collection}
-          onCollectionChange={(collection) => updateFilters({ collection })}
         />
         <Table
           stacked={breakpoints.sm}
